+++ paths.sh
@@ -8,6 +8,8 @@ TPS_EVAL_ROOT="$(dirname "$(realpath "${BASH_SOURCE[0]}")")"
 
 TPS_EVAL_ENV="tps_eval" # tps_eval conda environment name
 
+ESMFOLD_ENV="esmfold" # ESMFold (structure prediction) conda environment name
+
 SOLUPROT_PATH="/home2/soldat/documents/soluprot"
 SOLUPROT_ENV="soluprot" # SoluProt conda environment name